requirements.txt
Django==5.2.7
djangorestframework==3.14.0
ultralytics==8.3.41
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
PyMySQL==1.1.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3
